Feature/orders

+++ src/main/java/com/greatminds/ayni/shopping/domain/model/commands/CreateOrderCommand.java
@@ -2,6 +2,6 @@ package com.greatminds.ayni.shopping.domain.model.commands;
 
 import java.util.Date;
 
-public record CreateOrderCommand(String description, Double totalPrice, Integer quantity,  String paymentMethod, String status, Long saleId, Long orderedBy, Long acceptedBy, Date orderedDate) {
+public record CreateOrderCommand(String description, Double totalPrice, Long quantity,  String paymentMethod, String status, Long saleId, Long orderedBy, Long acceptedBy, Date orderedDate) {
 
 }
